Extract Looper to a separate file (#175)
